TOAMMEND added size to handlers and started sneaky builder

+++ src/common-stuff/math/pgcd.c
@@ -5,7 +5,7 @@
 
 
 //https://lemire.me/blog/2013/12/26/fastest-way-to-compute-the-greatest-common-divisor/
-unsigned int gcd(unsigned int u, unsigned int v)
+size_t gcd(size_t u, size_t v)
 {
 	int shift;
 	if (u == 0) return v;
@@ -15,7 +15,7 @@ unsigned int gcd(unsigned int u, unsigned int v)
 	do {
 		v >>= __builtin_ctz(v);
 		if (u > v) {
-			unsigned int t = v;
+			size_t t = v;
 			v = u;
 			u = t;
 		}  
@@ -25,9 +25,9 @@ unsigned int gcd(unsigned int u, unsigned int v)
 }
 
 typedef struct Conflict_element {
-	unsigned int value;
+	size_t value;
 	size_t origin;
-	unsigned int conflict_number;
+	size_t conflict_number;
 } Conflict_element;
 
 
@@ -36,12 +36,12 @@ static int cmpfunc ( const void * a, const void * b) {
 }
 
 
-static void prepare_conflict_table(unsigned int* numbers, size_t numbers_size, Conflict_element* res){
+static void prepare_conflict_table(size_t* numbers, size_t numbers_size, Conflict_element* res){
 	bzero((void*) res, numbers_size * sizeof(Conflict_element));
-	for (unsigned int i=0 ; i<numbers_size; i++){
+	for (int i=0 ; i<numbers_size; i++){
 		res[i].origin=i;
 		res[i].value= numbers[i];
-		for (unsigned int j=i+1 ; j<numbers_size;j++){
+		for (int j=i+1 ; j<numbers_size;j++){
 			if (gcd(numbers[i],numbers[j]) !=1){
 				res[i].conflict_number++;
 				res[j].conflict_number++;
@@ -53,12 +53,12 @@ static void prepare_conflict_table(unsigned int* numbers, size_t numbers_size, C
 }
 
 //Room for improvement. This algo is fast but sensible to bad input(eg 2 facto). Also it does not care to increment small value, which is bad for our use case
-unsigned int make_prime_together(unsigned int* numbers, size_t numbers_size){
+size_t make_prime_together(size_t* numbers, size_t numbers_size){
 	Conflict_element conflict_table [numbers_size]; 
 	prepare_conflict_table(numbers, numbers_size, conflict_table);
 
-	for (unsigned int i=0 ; i<numbers_size; i++){
-		unsigned int j=0;
+	for (int i=0 ; i<numbers_size; i++){
+		int j=0;
 		while (j<numbers_size ){
 			if (j==i){
 				j++;
@@ -75,9 +75,9 @@ unsigned int make_prime_together(unsigned int* numbers, size_t numbers_size){
 	}
 }
 
-bool prime_together(unsigned int* numbers, size_t size){
-	for (unsigned int i=0 ; i<size; i++)
-		for (unsigned int j=i+1 ; j<size;j++){
+bool prime_together(size_t* numbers, size_t size){
+	for (int i=0 ; i<size; i++)
+		for (int j=i+1 ; j<size;j++){
 			if (gcd(numbers[i],numbers[j]) !=1){
 				return false;
 			}
