+++ src/common-stuff/math/pgcd.h
@@ -1,11 +1,11 @@
 #include <stdbool.h>
 #include <stddef.h>
-unsigned int gcd(unsigned int u, unsigned int v);
+size_t gcd(size_t u, size_t v);
 //compute gcd of two numbers
 
-unsigned int make_prime_together(unsigned int* numbers, size_t size);
+size_t make_prime_together(size_t* numbers, size_t size);
 //take a list of number and try to make them prime one to each other with minimimum modification
 
-bool prime_together(unsigned int* numbers, size_t size);
+bool prime_together(size_t* numbers, size_t size);
 //Return true if numbers are prime on toward each other
